requirements.txt
pyyaml==5.4.1
scikit-learn
matplotlib
keras==2.4.3
tensorflow
pandas
matplotlib
numpy
opt-einsum
gast
astunparse
flatbuffers
keras_preprocessing
h5py